:neckbeard: more tests in Rust moved Rust tests to separate module
